# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "API Minimarket",
                 "description":  "DocumentaciÃ³n de la API del sistema Minimarket",
                 "version":  "1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
...
    "paths":  {
...
                  "/api/productos":  {
                                         "get":  {
                                                     "tags":  [
                                                                  "Producto"
                                                              ],
                                                     "summary":  "Obtener lista de productos",
                                                     "description":  "Devuelve una lista paginada de todos los productos disponibles en el sistema. Requiere autenticaciÃ³n.",
                                                     "operationId":  "listarTodos",
                                                     "parameters":  [
                                                                        {
                                                                            "name":  "page",
                                                                            "in":  "query",
                                                                            "required":  false,
                                                                            "schema":  {
                                                                                           "type":  "integer",
                                                                                           "format":  "int32",
                                                                                           "default":  0
                                                                                       }
                                                                        },
                                                                        {
                                                                            "name":  "size",
                                                                            "in":  "query",
                                                                            "required":  false,
                                                                            "schema":  {
                                                                                           "type":  "integer",
                                                                                           "format":  "int32",
                                                                                           "default":  10
                                                                                       }
                                                                        },
                                                                        {
                                                                            "name":  "sortBy",
                                                                            "in":  "query",
                                                                            "required":  false,
                                                                            "schema":  {
                                                                                           "type":  "string",
                                                                                           "default":  "nombre"
                                                                                       }
                                                                        },
                                                                        {
                                                                            "name":  "sortDir",
                                                                            "in":  "query",
                                                                            "required":  false,
                                                                            "schema":  {
                                                                                           "type":  "string",
                                                                                           "default":  "asc"
                                                                                       }
                                                                        }
                                                                    ],
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "Lista de productos obtenida exitosamente",
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/ProductoPageResponseDTO"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               }
                                                                   },
                                                     "security":  [
                                                                      {
                                                                          "bearerAuth":  [

                                                                                         ]
                                                                      }
                                                                  ]
                                                 },
                                         "post":  {
                                                      "tags":  [
                                                                   "Producto"
                                                               ],
                                                      "summary":  "Agregar un nuevo producto",
                                                      "description":  "Crea un nuevo producto en el sistema y devuelve los detalles del producto creado.",
                                                      "operationId":  "crearProducto",
                                                      "requestBody":  {
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/Producto"
                                                                                                                              }
                                                                                                               }
                                                                                      },
                                                                          "required":  true
                                                                      },
                                                      "responses":  {
                                                                        "201":  {
                                                                                    "description":  "Producto creado exitosamente",
                                                                                    "content":  {
                                                                                                    "*/*":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/EntityModelProductoResponseDTO"
                                                                                                                           }
                                                                                                            }
                                                                                                }
                                                                                }
                                                                    },
                                                      "security":  [
                                                                       {
                                                                           "bearerAuth":  [

                                                                                          ]
                                                                       }
                                                                   ]
                                                  }
                                     },
...
                  "/api/inventario/producto/{productoId}":  {
                                                                "get":  {
                                                                            "tags":  [
                                                                                         "Inventario"
                                                                                     ],
                                                                            "summary":  "Obtener movimientos por producto",
                                                                            "description":  "Devuelve los movimientos de un producto especÃ­fico",
                                                                            "operationId":  "listarPorProducto",
                                                                            "parameters":  [
                                                                                               {
                                                                                                   "name":  "productoId",
                                                                                                   "in":  "path",
                                                                                                   "required":  true,
                                                                                                   "schema":  {
                                                                                                                  "type":  "integer",
                                                                                                                  "format":  "int64"
                                                                                                              }
                                                                                               }
                                                                                           ],
                                                                            "responses":  {
                                                                                              "200":  {
                                                                                                          "description":  "Movimientos obtenidos exitosamente",
                                                                                                          "content":  {
                                                                                                                          "*/*":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/CollectionModelEntityModelInventario"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                                      }
                                                                                                      }
                                                                                          },
                                                                            "security":  [
                                                                                             {
                                                                                                 "bearerAuth":  [

                                                                                                                ]
                                                                                             }
                                                                                         ]
                                                                        }
                                                            },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "Producto":  {
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "id":  {
                                                                                      "type":  "integer",
                                                                                      "format":  "int64"
                                                                                  },
                                                                           "nombre":  {
                                                                                          "type":  "string"
                                                                                      },
                                                                           "precio":  {
                                                                                          "type":  "number",
                                                                                          "format":  "double"
                                                                                      },
                                                                           "stock":  {
                                                                                         "type":  "integer",
                                                                                         "format":  "int32"
                                                                                     },
                                                                           "categoria":  {
                                                                                             "$ref":  "#/components/schemas/Categoria"
                                                                                         }
                                                                       }
                                                    },
...
                                       "Embedded":  {
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "productoList":  {
                                                                                                "type":  "array",
                                                                                                "items":  {
                                                                                                              "$ref":  "#/components/schemas/ProductoResponseDTO"
                                                                                                          }
                                                                                            }
                                                                       }
                                                    },
                                       "ProductoPageResponseDTO":  {
                                                                       "type":  "object",
                                                                       "description":  "DTO para la respuesta paginada de productos en formato HATEOAS",
                                                                       "properties":  {
                                                                                          "page":  {
                                                                                                       "$ref":  "#/components/schemas/PageMetadata",
                                                                                                       "description":  "Metadatos de la pÃ¡gina actual"
                                                                                                   },
                                                                                          "_embedded":  {
                                                                                                            "$ref":  "#/components/schemas/Embedded",
                                                                                                            "description":  "Contenedor de la lista de productos"
                                                                                                        },
                                                                                          "_links":  {
                                                                                                         "type":  "object",
                                                                                                         "additionalProperties":  {
                                                                                                                                      "$ref":  "#/components/schemas/Link"
                                                                                                                                  },
                                                                                                         "description":  "Enlaces de navegaciÃ³n de la paginaciÃ³n"
                                                                                                     }
                                                                                      }
                                                                   },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "description":  "AutenticaciÃ³n JWT usando Bearer Token",
                                                                  "in":  "header",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}